fix: renaming to lowercase
